Implement GBC rendering, and refactor codebase to improve modularity.
This commit implements the entirety of GBC rendering logic, including
fallback to DMG mode for non-gbc-supported roms. This commit also
includes a somewhat major refactor to ppu render logic, and a complete
overhaul of the project layout. The core gb logic has been separated
into a seperate package from the window and key handling logic.

This commit has been tested with:

- link's awakening dx
- pokemon yellow
- pokemon silver
- donkey kong country
- all previously-tested gb games

This commit hangs on the intro screen for oracle of ages. Initial
testing indicates that this is due to an issue with sound register
read values.